// src/generated/model/intakeRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { IntakeRequestStatus } from './intakeRequestStatus.ts';

export interface IntakeRequest {
  academicYearId: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  startDate: string;
  endDate: string;
  /** @minimum 1 */
  totalCapacity?: number;
  status: IntakeRequestStatus;
  remarks?: string;
}