// src/generated/model/updateIntakeRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateIntakeRequestStatus } from './updateIntakeRequestStatus.ts';

export interface UpdateIntakeRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  startDate: string;
  endDate: string;
  /** @minimum 1 */
  totalCapacity?: number;
  status: UpdateIntakeRequestStatus;
  remarks?: string;
}