// src/generated/model/intakeResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { IntakeResponseStatus } from './intakeResponseStatus.ts';

export interface IntakeResponse {
  id?: string;
  academicYearId?: string;
  code?: string;
  name?: string;
  startDate?: string;
  endDate?: string;
  totalCapacity?: number;
  status?: IntakeResponseStatus;
  remarks?: string;
  createdAt?: string;
  updatedAt?: string;
}